% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.tree.R
\name{addAlignment}
\alias{addAlignment}
\title{add alignments to list}
\usage{
addAlignment(alist, fileName)
}
\arguments{
\item{alist}{(possibly empty) list of alignments in DNAbin format}

\item{fileName}{fasta file name}
}
\value{
list of alignments
}
\description{
add alignments to list
}
