% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_tree.R
\name{createMLTrees}
\alias{createMLTrees}
\title{create maximum likelihood trees from a list of alignments}
\usage{
createMLTrees(alist, maxit = 10000, optNni = TRUE, bf = basefreqs,
  untangle = TRUE, ...)
}
\arguments{
\item{alist}{list of alignments in DNAbin format}

\item{maxit}{Max number of iterations in optimisation}

\item{optNni}{(TRUE) optimise the topology of the tree}

\item{untangle}{flag as to whether to untangle tree, may be needed for small trees}

\item{...}{further parameters for optim.pml}
}
\value{
list of maximum likelihood trees in phylo format
}
\description{
create maximum likelihood trees from a list of alignments
}
