% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_tree.R
\name{createTimedTrees}
\alias{createTimedTrees}
\title{create timed trees from a list of maximum likelihood trees}
\usage{
createTimedTrees(mltrees, sampledates, alist, strictClock = T,
  meanRateLimits = c(0.3, 0.7))
}
\arguments{
\item{mltrees}{list of trees as created by createMLTrees}

\item{sampledates}{vector of sample dates}

\item{alist}{list of alignments in DNAbin format}

\item{meanRateLimits}{vector of length 2 specifying upper and lower rate limits for treedater}

\item{(TRUE)}{set whether to use strict clock in treedater}
}
\value{
list of timed trees in phylo format
}
\description{
create timed trees from a list of maximum likelihood trees
}
