% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.tree.R
\name{estimate.tree}
\alias{estimate.tree}
\title{estimate a tree}
\usage{
estimate.tree(aln, maxit = 10000, addCols = TRUE, optNni = TRUE,
  bf = basefreqs, ...)
}
\arguments{
\item{aln}{Alignment in DNAbin format}

\item{maxit}{Max number of iterations in optimisation}

\item{optNni}{(TRUE) optimise the topology of the tree}

\item{...}{further parameters for optim.pml}
}
\description{
estimate a tree
}
