% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_tree.R
\name{estimatetree}
\alias{estimatetree}
\title{estimate a tree}
\usage{
estimatetree(aln, maxit = 10000, optNni = TRUE, bf = basefreqs, ...)
}
\arguments{
\item{aln}{Alignment in DNAbin format}

\item{maxit}{Max number of iterations in optimisation}

\item{optNni}{(TRUE) optimise the topology of the tree}

\item{...}{further parameters for optim.pml}
}
\value{
tree object of type phylo
}
\description{
estimate a tree
}
\examples{
test=estimatetree(aln)
myModel <- setDatesFromFile(myModel, datefile)
}
