% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInfectionTimes.R
\name{filterPossibleWiw}
\alias{filterPossibleWiw}
\title{Filter record on specified allowable transmission matrix}
\usage{
filterPossibleWiw(record, pMatrix)
}
\arguments{
\item{record}{MCMC output produced by inferTTree}

\item{pMatrix}{Matrix of allowable transmissions, TRUE for allowed, FALSE for not allowed}
}
\value{
Filtered record
}
\description{
Filter record on specified allowable transmission matrix
}
