% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInfectionTimes.R
\name{filterTransPair}
\alias{filterTransPair}
\title{Filter record on specified directed transmission pair}
\usage{
filterTransPair(record, from, to)
}
\arguments{
\item{record}{MCMC output produced by inferTTree}

\item{from}{From case index, either an integer or a string matching one of the case names in the data}

\item{to}{To case index, either an integer or a string matching one of the case names in the data}
}
\value{
Filtered record
}
\description{
Filter record on specified directed transmission pair
}
