% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transphylo_extras.R
\name{getGenerationTimes}
\alias{getGenerationTimes}
\title{Get generation times}
\usage{
getGenerationTimes(ctree)
}
\arguments{
\item{ctree}{Combined tree object 'ctree' from the record produced by inferTTree, for example record[[1]]$ctree where record is the output of inferTTree}
}
\value{
Vector of times between becoming infected and infecting others (generation times) in the posterior
}
\description{
Get generation times
}
\examples{
getGenerationTimes(record[[1]]$ctree)
}
