% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transphylo_extras.R
\name{getInfectionDates}
\alias{getInfectionDates}
\title{Get infection dates}
\usage{
getInfectionDates(ctree)
}
\arguments{
\item{ctree}{TransPhylo tree}
}
\value{
A vector of posterior infection dates
}
\description{
Get infection dates
}
\examples{
getInfectionDates(ctree)
}
