% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{getInfectionTimes}
\alias{getInfectionTimes}
\title{Get infection times}
\usage{
getInfectionTimes(record, k)
}
\arguments{
\item{record}{MCMC output produced by inferTTree}

\item{k}{Case whose posterior infection times are to be extracted. Either an integer or a string matching one of the case names in the data}
}
\value{
A vector of posterior infection times for case k
}
\description{
Get infection times
}
\examples{
getInfectionTimes(record)
}
