% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getLikelyInfectors}
\alias{getLikelyInfectors}
\title{Get infection likelihoods for a given case}
\usage{
getLikelyInfectors(record, case, burnin = 0.5)
}
\arguments{
\item{record}{MCMC output produced by inferTTree}

\item{case}{Index or label can be supplied}

\item{burnin}{Proportion of record entries to use}
}
\description{
Get infection likelihoods for a given case
}
