% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transphylo_extras.R
\name{getNumberUnsampled}
\alias{getNumberUnsampled}
\title{Get the number of unsampled cases}
\usage{
getNumberUnsampled(ctree)
}
\arguments{
\item{ctree}{Combined tree object 'ctree' from the record produced by inferTTree, for example record[[1]]$ctree where record is the output of inferTTree}
}
\value{
The number of unsampled cases in ctree
}
\description{
Get the number of unsampled cases
}
\examples{
getNumberUnsampled(ctree)
}
