% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getTimesToSampling}
\alias{getTimesToSampling}
\title{Get infection-to-sampling times}
\usage{
getTimesToSampling(ctree)
}
\arguments{
\item{ctree}{TransPhylo tree}
}
\value{
A vector of posterior times to sampling
}
\description{
Get infection-to-sampling times
}
\examples{
getTimesToSampling(ctree)
}
