% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getTTreeDistInfo}
\alias{getTTreeDistInfo}
\title{Create list of wiw information in order to compute transmission tree distances}
\usage{
getTTreeDistInfo(record, skipnum = 1)
}
\arguments{
\item{record}{MCMC output produced by inferTTree}

\item{skipnum}{Number of record entries to skip, ie skipnum=10 uses the 1st, 11th, 21st ... elements of the record}
}
\value{
list of MRCI information required by wiwTreeDist, one entry for each transmission tree that is included
}
\description{
Create list of wiw information in order to compute transmission tree distances
}
