% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transphylo_extras.R
\name{networkPlot}
\alias{networkPlot}
\title{Simple visualisation of transmission tree with visNetwork}
\usage{
networkPlot(ctree)
}
\arguments{
\item{ctree}{Combined tree object 'ctree' from the record produced by inferTTree, for example record[[1]]$ctree where record is the output of inferTTree}
}
\description{
Simple visualisation of transmission tree with visNetwork
}
