% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{plot_times_to_samp}
\alias{plot_times_to_samp}
\title{Plot histogram of times to sampling for each cluster}
\usage{
plot_times_to_samp(cls_record, burnin = 0.5, ...)
}
\arguments{
\item{cls_record}{Posterior MCMC sample set from TransPhylo for all clusters}

\item{burnin}{Proportion of the MCMC output to be discarded as burnin}

\item{...}{parameters to be passed to grid.arrange}
}
\description{
Plot histogram of times to sampling for each cluster
}
