% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{plot_unsampled_cases}
\alias{plot_unsampled_cases}
\title{Plot histogram of the number of unsampled cases for each cluster}
\usage{
plot_unsampled_cases(cls_record, burnin = 0.5, ...)
}
\arguments{
\item{cls_record}{Posterior MCMC sample set from TransPhylo for all clusters}

\item{burnin}{Proportion of the MCMC output to be discarded as burnin}

\item{...}{parameters to be passed to grid.arrange}
}
\description{
Plot histogram of the number of unsampled cases for each cluster
}
