% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{plotGenerationTimeDensity}
\alias{plotGenerationTimeDensity}
\title{Plot generation time density for chosen case}
\usage{
plotGenerationTimeDensity(singleRecord, index, burnin = 0.5)
}
\arguments{
\item{singleRecord}{Posterior MCMC sample set from TransPhylo}

\item{index}{Index of case}

\item{burnin}{Proportion of the MCMC output to be discarded as burnin}
}
\value{
A vector of posterior generation times
}
\description{
Plot generation time density for chosen case
}
\examples{
plotGenerationTimeDensity(record, index)
}
