% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{plotSampleTimeDensity}
\alias{plotSampleTimeDensity}
\title{Plot time to sampling (from infection) for chosen case}
\usage{
plotSampleTimeDensity(singleRecord, index)
}
\arguments{
\item{singleRecord}{Posterior sample set of TransPhylo trees}

\item{index}{Index of case}
}
\value{
A vector of posterior times
}
\description{
Plot time to sampling (from infection) for chosen case
}
\examples{
plotSampleTimeDensity(record, index)
}
