% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{plotTransTreeSummary}
\alias{plotTransTreeSummary}
\title{Plot transmission network with posterior probabilities at or above given cutoff}
\usage{
plotTransTreeSummary(record, cutoff = 1, burnin = 0.5,
  nodeColour = "lightblue", fontSize = 24)
}
\arguments{
\item{record}{Posterior MCMC sample set from TransPhylo}

\item{cutoff}{Minimum probabilty for edge to be included}

\item{burnin}{Proportion of the MCMC output to be discarded as burnin}

\item{nodeColour}{Colour for all nodes}

\item{fontSize}{Font side for node labels}
}
\description{
Plot transmission network with posterior probabilities at or above given cutoff
}
