% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{setUpDates}
\alias{setUpDates}
\title{create list of dates to match alignments}
\usage{
setUpDates(alist, allDates)
}
\arguments{
\item{alist}{(possibly empty) list of alignments in DNAbin format}

\item{allDates}{data frame containg all IDs and dates as read from a csv file}
}
\value{
list of dates
}
\description{
create list of dates to match alignments
}
