% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.tree.R
\name{simulTransTrees}
\alias{simulTransTrees}
\title{Simultaneously infer transmission trees given phylogenetic trees constructed from clusters of sequences.
Wrapper for infer_multiTTree_shareParam, taking mean and std deviations rather than shape and scale for the Gamma parameters}
\usage{
simulTransTrees(timedTrees, mean_gen = 1, stddev_gen = 1, mean_sample = 1,
  stddev_sample = 1, share = c("neg", "off.r", "off.p", "pi"),
  mcmcIterations = nIter, startNeg = 1, startOff.p = 0.8,
  startOff.r = 1, startPi = 0.9, updateNeg = T, updatePi = F,
  updateOff.p = FALSE)
}
\arguments{
\item{mean_gen}{Mean of the Gamma probability density function representing the generation time}

\item{stddev_gen}{Standard deviation of the Gamma probability density function representing the generation time}

\item{mean_sample}{Mean of the Gamma probability density function representing the sampling time}

\item{stddev_sample}{Standard deviation of the Gamma probability density function representing the sampling time}

\item{share}{Character vector of parameters to be shared. For example, share = c("off.r", "off.p") would
share the offspring distribution. Allowed parameter names are "neg", "off.r", "off.p" and "pi".}

\item{mcmcIterations}{Number of MCMC iterations to run the algorithm for}

\item{startNeg}{Starting value of within-host coalescent parameter Ne*g}

\item{startOff.p}{Starting value of parameter off.p}

\item{startOff.r}{Starting value of parameter off.r}

\item{startPi}{Starting value of sampling proportion pi}

\item{updateNeg}{Whether of not to update the parameter Ne*g}

\item{updatePi}{Whether or not to update the parameter pi}

\item{updateOff.p}{Whether or not to update the parameter off.p}

\item{ptree_lst}{List of phylogenetic tree}

\item{thinning}{MCMC thinning interval between two sampled iterations}

\item{updateOff.r}{Whether or not to update the parameter off.r}

\item{startCTree_lst}{Optional combined list of trees to start from}

\item{updateTTree}{Whether or not to update the transmission tree}

\item{optiStart}{Whether or not to optimise the MCMC start point}

\item{dateT}{Date when process stops (this can be Inf for fully simulated outbreaks)}
}
\value{
posterior sample set of transmission trees for all clusters
}
\description{
Simultaneously infer transmission trees given phylogenetic trees constructed from clusters of sequences.
Wrapper for infer_multiTTree_shareParam, taking mean and std deviations rather than shape and scale for the Gamma parameters
}
