% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transphylo_extras.R
\name{getTimesToSampling}
\alias{getTimesToSampling}
\title{Get infection times}
\usage{
getTimesToSampling(ctree)
}
\arguments{
\item{ctree}{TransPhylo tree}
}
\value{
A vector of posterior times to sampling
}
\description{
Get infection times
}
\examples{
getTimesToSampling(ttree)
}
