% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transphylo_extras.R
\name{networkTPlot}
\alias{networkTPlot}
\title{dynamic plot of the transmission network with edges weighted according  to likelihood of transmission}
\usage{
networkTPlot(thisRecord, mcmcIndex = 1, missLabel = "Unsampled",
  colours = c("lightblue", "orange"))
}
\arguments{
\item{thisRecord}{Posterior sample set of transmission trees for all clusters}

\item{mcmcIndex}{Sample set index of network to be displayed}

\item{missLabel}{Label to be used for missing cases}

\item{colours}{Colours for sampled and unsampled nodes, vector of 2 colours}
}
\description{
dynamic plot of the transmission network with edges weighted according  to likelihood of transmission
}
\examples{
networkTPlot(record)
}
