% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transphylo_extras.R
\name{plot_times_to_samp}
\alias{plot_times_to_samp}
\title{Plot histogram of times to sampling for each cluster}
\usage{
plot_times_to_samp(cls_record, ...)
}
\arguments{
\item{cls_record}{Posterior sample set of transmission trees for all clusters}

\item{...}{parameters to be passed to grid.arrange}
}
\description{
Plot histogram of times to sampling for each cluster
}
