% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{plot_unsampled_cases}
\alias{plot_unsampled_cases}
\title{Plot histogram of the number of unsampled cases for each cluster}
\usage{
plot_unsampled_cases(cls_record, ...)
}
\arguments{
\item{cls_record}{Posterior sample set of TransPhylo trees for all clusters}
}
\description{
Plot histogram of the number of unsampled cases for each cluster
}
