% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{plotGenerationTimeDensity}
\alias{plotGenerationTimeDensity}
\title{Plot generation time density for chosen case}
\usage{
plotGenerationTimeDensity(singleRecord, index)
}
\arguments{
\item{singleRecord}{Posterior sample set of TransPhylo trees}

\item{index}{Index of case}
}
\value{
A vector of posterior generation times
}
\description{
Plot generation time density for chosen case
}
\examples{
plotGenerationTimeDensity(record, index)
}
