% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{plotInfectionDateDensity}
\alias{plotInfectionDateDensity}
\title{Plot infection date density for chosen case}
\usage{
plotInfectionDateDensity(singleRecord, index, overlayDate)
}
\arguments{
\item{singleRecord}{Posterior sample set of TransPhylo trees}

\item{index}{Index of case}

\item{overlayDate}{Date to overlay on graph}
}
\value{
A vector of posterior infection dates
}
\description{
Plot infection date density for chosen case
}
\examples{
plotInfectionDateDensity(record, index, date)
}
