% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visuals.R
\name{plotSampleTimeDensity}
\alias{plotSampleTimeDensity}
\title{Plot time to sampling (from infection) for chosen case}
\usage{
plotSampleTimeDensity(singleRecord, index, burnin = 0.5)
}
\arguments{
\item{singleRecord}{Posterior MCMC sample set from TransPhylo}

\item{index}{Index of case}

\item{burnin}{Proportion of the MCMC output to be discarded as burnin}
}
\value{
A vector of posterior times
}
\description{
Plot time to sampling (from infection) for chosen case
}
\examples{
plotSampleTimeDensity(record, index)
}
