% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transphylo_extras.R
\name{plotTransTreeSummary}
\alias{plotTransTreeSummary}
\title{Plot transmission network with posterior probabilities at or above given cutoff}
\usage{
plotTransTreeSummary(record, cutoff = 1, burnin = 0.5,
  nodeColour = "lightblue", fontSize = 24, edgeWidth = 1)
}
\arguments{
\item{record}{MCMC output produced by inferTTree}

\item{cutoff}{Minimum distance to be included}

\item{burnin}{Proportion of record entries to use}
}
\description{
Plot transmission network with posterior probabilities at or above given cutoff
}
